refactor(hub): split middleware.py into 3 files under 300-line limit
CI file-length check requires max 300 lines per file. Split:
- middleware.py (142 lines) — protocol, context, runner, factory
- middleware_stages.py (244 lines) — stages 1-5 (guards + command)
- middleware_submit.py (244 lines) — stage 6 (pool submit + resume)
